+++ src/ComputeFleet/ComputeFleet.Autorest/test/New-AzComputeFleet.Tests.ps1
@@ -28,10 +28,15 @@ Describe 'New-AzComputeFleet' {
         $managedFleetJsonName = "managed-fleet-json"
         $managedFleetJsonStrName = "managed-fleet-jsonstr"
 
-        $result = New-TestResourceGroup -ResourceGroupName $resourceGroupName `
-            -Location $env.Location -VNetName $vnetName -NsgName $nsgName
-        $subnetId = $result.SubnetId
-        $nsgId = $result.NsgId
+        if ($TestMode -ne 'playback') {
+            $result = New-TestResourceGroup -ResourceGroupName $resourceGroupName `
+                -Location $env.Location -VNetName $vnetName -NsgName $nsgName
+            $subnetId = $result.SubnetId
+            $nsgId = $result.NsgId
+        } else {
+            $subnetId = "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test/providers/Microsoft.Network/virtualNetworks/vnet/subnets/subnet1"
+            $nsgId = "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test/providers/Microsoft.Network/networkSecurityGroups/nsg"
+        }
     }
 
     It 'CreateLaunchModeFleet' {
@@ -234,6 +239,8 @@ Describe 'New-AzComputeFleet' {
     }
 
     AfterAll {
-        Remove-AzResourceGroup -Name $resourceGroupName -ErrorAction SilentlyContinue -Confirm:$false
+        if ($TestMode -ne 'playback') {
+            Remove-AzResourceGroup -Name $resourceGroupName -ErrorAction SilentlyContinue -Confirm:$false
+        }
     }
 }
